+++ dds-agent-cmd/src/Options.h
@@ -12,6 +12,7 @@
 #include "version.h"
 #include "Res.h"
 #include "ProtocolCommands.h"
+#include "BOOSTHelper.h"
 //=============================================================================
 namespace bpo = boost::program_options;
 //=============================================================================
@@ -20,9 +21,12 @@ namespace dds
     enum class EAgentCmdType
     {
         UNKNOWN = -1,
-        GETLOG = 0
+        GETLOG = 0,
+        UPDATE_KEY
     };
-    std::map<EAgentCmdType, std::string> AgentCmdTypeCodeToString = { { EAgentCmdType::GETLOG, "getlog" } };
+    typedef std::map<EAgentCmdType, std::string> mapAgentCmdTypeCodes_t;
+    mapAgentCmdTypeCodes_t AgentCmdTypeCodeToString = { { EAgentCmdType::GETLOG, "getlog" },
+                                                        { EAgentCmdType::UPDATE_KEY, "update-key" } };
 
     // A custom streamer to help boost program options to convert string options to EAgentCmdType
     inline std::istream& operator>>(std::istream& _in, EAgentCmdType& _agentCmd)
@@ -31,6 +35,8 @@ namespace dds
         _in >> token;
         if (token == "getlog")
             _agentCmd = EAgentCmdType::GETLOG;
+        else if (token == "update-key")
+            _agentCmd = EAgentCmdType::UPDATE_KEY;
         else
             throw bpo::invalid_option_value(token);
         return _in;
@@ -38,14 +44,7 @@ namespace dds
 
     inline std::ostream& operator<<(std::ostream& _out, EAgentCmdType& _agentCmd)
     {
-        switch (_agentCmd)
-        {
-            case EAgentCmdType::GETLOG:
-                _out << "getlog";
-                break;
-            case EAgentCmdType::UNKNOWN:
-                break;
-        }
+        _out << AgentCmdTypeCodeToString[_agentCmd];
         return _out;
     }
 
@@ -60,6 +59,8 @@ namespace dds
 
         bool m_sendCommandToAllAgents;
         EAgentCmdType m_agentCmd;
+        std::string m_sUpdKey_key;
+        std::string m_sUpdKey_value;
     } SOptions_t;
     //=============================================================================
     inline void PrintVersion()
@@ -79,17 +80,30 @@ namespace dds
         bpo::options_description options("dds-agent-cmd options");
         options.add_options()("help,h", "Produce help message");
         options.add_options()("version,v", "Version information");
-        options.add_options()("cmd,c", bpo::value<EAgentCmdType>(&_options->m_agentCmd),
-                              "Command that has to be send to agents.");
+        options.add_options()(
+            "command", bpo::value<EAgentCmdType>(&_options->m_agentCmd),
+            "The command is a name of a dds-agent-cmd command."
+            " Can be one of the following: getlog, and update-key.\n"
+            "For user's convenience it is allowed to call dds-agent-cmd without \"--command\" option"
+            " by just specifying the command name directly, like:\ndds-agent-cmd getlog\n\n"
+            "Commands:\n"
+            "   getlog: \tRetrieve log files from worker nodes. Files will be saved in ~/.DDS/log/agents\n"
+            "   update-key: \tIt forces an update of a given task's property in the topology. Name of the property and "
+            "its value should be provided additionally (see --key and --value) \n");
         options.add_options()("all,a", "Send command to all active agents");
+        options.add_options()("key", bpo::value<std::string>(&_options->m_sUpdKey_key), "Specefies the key to update");
+        options.add_options()("value", bpo::value<std::string>(&_options->m_sUpdKey_value),
+                              "Specefies the new value ofthe given key");
 
         bpo::positional_options_description positional;
-        positional.add("cmd", -1);
+        positional.add("command", -1);
 
         // Parsing command-line
         bpo::variables_map vm;
         bpo::store(bpo::command_line_parser(_argc, _argv).options(options).positional(positional).run(), vm);
         bpo::notify(vm);
+
+        MiscCommon::BOOSTHelper::option_dependency(vm, "key", "value");
 
         if (vm.count("help") || vm.empty())
         {
@@ -101,9 +115,15 @@ namespace dds
             PrintVersion();
             return false;
         }
-        if (!vm.count("cmd"))
+        if (!vm.count("command"))
         {
-            LOG(MiscCommon::log_stderr) << "specify an agent command"
+            LOG(MiscCommon::log_stderr) << "Nothing to do. Please, specify a command"
+                                        << "\n\n" << options;
+            return false;
+        }
+        if (EAgentCmdType::UPDATE_KEY == _options->m_agentCmd && !vm.count("key"))
+        {
+            LOG(MiscCommon::log_stderr) << "Please, specify the key to update"
                                         << "\n\n" << options;
             return false;
         }